Move getLatestWALSnap into etcdutl/common.go
