+++ etcdutl/etcdutl/common.go
@@ -15,11 +15,18 @@
 package etcdutl
 
 import (
+	"errors"
+
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
 
 	"go.etcd.io/etcd/client/pkg/v3/logutil"
 	"go.etcd.io/etcd/pkg/v3/cobrautl"
+	"go.etcd.io/etcd/server/v3/etcdserver/api/snap"
+	"go.etcd.io/etcd/server/v3/storage/datadir"
+	"go.etcd.io/etcd/server/v3/storage/wal"
+	"go.etcd.io/etcd/server/v3/storage/wal/walpb"
+	"go.etcd.io/raft/v3/raftpb"
 )
 
 func GetLogger() *zap.Logger {
@@ -31,4 +38,33 @@ func GetLogger() *zap.Logger {
 		cobrautl.ExitWithError(cobrautl.ExitBadArgs, err)
 	}
 	return lg
+}
+
+func getLatestWALSnap(lg *zap.Logger, dataDir string) (walpb.Snapshot, error) {
+	snapshot, err := getLatestV2Snapshot(lg, dataDir)
+	if err != nil {
+		return walpb.Snapshot{}, err
+	}
+
+	var walsnap walpb.Snapshot
+	if snapshot != nil {
+		walsnap.Index, walsnap.Term = snapshot.Metadata.Index, snapshot.Metadata.Term
+	}
+	return walsnap, nil
+}
+
+func getLatestV2Snapshot(lg *zap.Logger, dataDir string) (*raftpb.Snapshot, error) {
+	walPath := datadir.ToWALDir(dataDir)
+	walSnaps, err := wal.ValidSnapshotEntries(lg, walPath)
+	if err != nil {
+		return nil, err
+	}
+
+	ss := snap.New(lg, datadir.ToSnapDir(dataDir))
+	snapshot, err := ss.LoadNewestAvailable(walSnaps)
+	if err != nil && !errors.Is(err, snap.ErrNoSnapshot) {
+		return nil, err
+	}
+
+	return snapshot, nil
 }
